pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
[project]
name = "PyKrasue"
version = "0.1.7"
authors = [
    {name="Andrew Mengede", email="amengede27@outlook.com"},
]
description = "A GPU accelerated pygame/arcade alternative"
readme = "readme.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = ["pygame-ce", "PyOpenGL", "Pillow", "numpy"]

[project.urls]
Homepage = "https://github.com/amengede/pykrasue"
Issues = "https://github.com/amengede/pykrasue/issues"